modification: removed testAlarm, corrected setting of currentAlarm
The redundant testAlarm removed and replaced by currentAlarm.
The current alarm is set only inside the function setCurrentAlarm. The current alarm audio was set before when clicked play button - it wasn't the best solution (during debugging or testing it was confusing what is the current alarm audio).
Now, the current alarm is set (by calling setCurrentAlarm()) when:
    - selected other option inside HTML Select with alarm names in the modal EDIT
    - closing modal EDIT without saving - current alarm will be reset from Local Storage stored value
Also corrected loading of current alarm audio. For example, when current alarm is previewed (played) inside the modal EDIT and then saved, it needs to load current alarm audio from time 0 to play from the start when alarm is triggered on main page of the timer.